// frontend/src/app/core/api/model/apiResponseAdminMeResponse.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminMeResponseDto } from './adminMeResponse';


export interface ApiResponseAdminMeResponseDto { 
    success?: boolean;
    status?: number;
    message?: string;
    data?: AdminMeResponseDto;
    errors?: { [key: string]: string; };
    timestamp?: string;
}