// frontend/src/app/core/api/model/apiResponseAuthResponse.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AuthResponseDto } from './authResponse';


export interface ApiResponseAuthResponseDto { 
    success?: boolean;
    status?: number;
    message?: string;
    data?: AuthResponseDto;
    errors?: { [key: string]: string; };
    timestamp?: string;
}